requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pymongo>=4.5.0
paho-mqtt>=2.0.0
networkx>=3.1
numpy>=1.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
openai>=1.0.0
